DAS-2067 - Migrate datatree/tests/test_dataset_api.py.
